refactor: 바뀐 의존성에 따른 리팩토링
